KODI64: General Kodi Code x64 fixes

+++ lib/cpluff/libcpluff/win32/dirent.c
@@ -21,7 +21,7 @@ extern "C"
 
 struct DIR
 {
-    long                handle; /* -1 for failed rewind */
+    intptr_t            handle; /* -1 for failed rewind */
     struct _finddata_t  info;
     struct dirent       result; /* d_name null iff first time */
     char                *name;  /* null-terminated char string */
